use the same app structure expected by the project generator
